+++ i18next.js
@@ -488,13 +488,13 @@
       const resolved = this.resolve(key, options);
       return resolved?.res !== undefined;
     }
-    extractFromKey(key, options) {
-      let nsSeparator = options.nsSeparator !== undefined ? options.nsSeparator : this.options.nsSeparator;
+    extractFromKey(key, opt) {
+      let nsSeparator = opt.nsSeparator !== undefined ? opt.nsSeparator : this.options.nsSeparator;
       if (nsSeparator === undefined) nsSeparator = ':';
-      const keySeparator = options.keySeparator !== undefined ? options.keySeparator : this.options.keySeparator;
-      let namespaces = options.ns || this.options.defaultNS || [];
+      const keySeparator = opt.keySeparator !== undefined ? opt.keySeparator : this.options.keySeparator;
+      let namespaces = opt.ns || this.options.defaultNS || [];
       const wouldCheckForNsInKey = nsSeparator && key.indexOf(nsSeparator) > -1;
-      const seemsNaturalLanguage = !this.options.userDefinedKeySeparator && !options.keySeparator && !this.options.userDefinedNsSeparator && !options.nsSeparator && !looksLikeObjectPath(key, nsSeparator, keySeparator);
+      const seemsNaturalLanguage = !this.options.userDefinedKeySeparator && !opt.keySeparator && !this.options.userDefinedNsSeparator && !opt.nsSeparator && !looksLikeObjectPath(key, nsSeparator, keySeparator);
       if (wouldCheckForNsInKey && !seemsNaturalLanguage) {
         const m = key.match(this.interpolator.nestingRegexp);
         if (m && m.length > 0) {
@@ -512,28 +512,28 @@
         namespaces: isString(namespaces) ? [namespaces] : namespaces
       };
     }
-    translate(keys, options, lastKey) {
-      if (typeof options !== 'object' && this.options.overloadTranslationOptionHandler) {
-        options = this.options.overloadTranslationOptionHandler(arguments);
+    translate(keys, opt, lastKey) {
+      if (typeof opt !== 'object' && this.options.overloadTranslationOptionHandler) {
+        opt = this.options.overloadTranslationOptionHandler(arguments);
       }
-      if (typeof options === 'object') options = {
-        ...options
+      if (typeof options === 'object') opt = {
+        ...opt
       };
-      if (!options) options = {};
+      if (!opt) opt = {};
       if (keys == null) return '';
       if (!Array.isArray(keys)) keys = [String(keys)];
-      const returnDetails = options.returnDetails !== undefined ? options.returnDetails : this.options.returnDetails;
-      const keySeparator = options.keySeparator !== undefined ? options.keySeparator : this.options.keySeparator;
+      const returnDetails = opt.returnDetails !== undefined ? opt.returnDetails : this.options.returnDetails;
+      const keySeparator = opt.keySeparator !== undefined ? opt.keySeparator : this.options.keySeparator;
       const {
         key,
         namespaces
-      } = this.extractFromKey(keys[keys.length - 1], options);
+      } = this.extractFromKey(keys[keys.length - 1], opt);
       const namespace = namespaces[namespaces.length - 1];
-      const lng = options.lng || this.language;
-      const appendNamespaceToCIMode = options.appendNamespaceToCIMode || this.options.appendNamespaceToCIMode;
+      const lng = opt.lng || this.language;
+      const appendNamespaceToCIMode = opt.appendNamespaceToCIMode || this.options.appendNamespaceToCIMode;
       if (lng?.toLowerCase() === 'cimode') {
         if (appendNamespaceToCIMode) {
-          const nsSeparator = options.nsSeparator || this.options.nsSeparator;
+          const nsSeparator = opt.nsSeparator || this.options.nsSeparator;
           if (returnDetails) {
             return {
               res: `${namespace}${nsSeparator}${key}`,
@@ -541,7 +541,7 @@
               exactUsedKey: key,
               usedLng: lng,
               usedNS: namespace,
-              usedParams: this.getUsedParamsDetails(options)
+              usedParams: this.getUsedParamsDetails(opt)
             };
           }
           return `${namespace}${nsSeparator}${key}`;
@@ -553,26 +553,26 @@
             exactUsedKey: key,
             usedLng: lng,
             usedNS: namespace,
-            usedParams: this.getUsedParamsDetails(options)
+            usedParams: this.getUsedParamsDetails(opt)
           };
         }
         return key;
       }
-      const resolved = this.resolve(keys, options);
+      const resolved = this.resolve(keys, opt);
       let res = resolved?.res;
       const resUsedKey = resolved?.usedKey || key;
       const resExactUsedKey = resolved?.exactUsedKey || key;
       const noObject = ['[object Number]', '[object Function]', '[object RegExp]'];
-      const joinArrays = options.joinArrays !== undefined ? options.joinArrays : this.options.joinArrays;
+      const joinArrays = opt.joinArrays !== undefined ? opt.joinArrays : this.options.joinArrays;
       const handleAsObjectInI18nFormat = !this.i18nFormat || this.i18nFormat.handleAsObject;
-      const needsPluralHandling = options.count !== undefined && !isString(options.count);
-      const hasDefaultValue = Translator.hasDefaultValue(options);
-      const defaultValueSuffix = needsPluralHandling ? this.pluralResolver.getSuffix(lng, options.count, options) : '';
-      const defaultValueSuffixOrdinalFallback = options.ordinal && needsPluralHandling ? this.pluralResolver.getSuffix(lng, options.count, {
+      const needsPluralHandling = opt.count !== undefined && !isString(opt.count);
+      const hasDefaultValue = Translator.hasDefaultValue(opt);
+      const defaultValueSuffix = needsPluralHandling ? this.pluralResolver.getSuffix(lng, opt.count, opt) : '';
+      const defaultValueSuffixOrdinalFallback = opt.ordinal && needsPluralHandling ? this.pluralResolver.getSuffix(lng, opt.count, {
         ordinal: false
       }) : '';
-      const needsZeroSuffixLookup = needsPluralHandling && !options.ordinal && options.count === 0;
-      const defaultValue = needsZeroSuffixLookup && options[`defaultValue${this.options.pluralSeparator}zero`] || options[`defaultValue${defaultValueSuffix}`] || options[`defaultValue${defaultValueSuffixOrdinalFallback}`] || options.defaultValue;
+      const needsZeroSuffixLookup = needsPluralHandling && !opt.ordinal && opt.count === 0;
+      const defaultValue = needsZeroSuffixLookup && opt[`defaultValue${this.options.pluralSeparator}zero`] || opt[`defaultValue${defaultValueSuffix}`] || opt[`defaultValue${defaultValueSuffixOrdinalFallback}`] || opt.defaultValue;
       let resForObjHndl = res;
       if (handleAsObjectInI18nFormat && !res && hasDefaultValue) {
         resForObjHndl = defaultValue;
@@ -580,17 +580,17 @@
       const handleAsObject = shouldHandleAsObject(resForObjHndl);
       const resType = Object.prototype.toString.apply(resForObjHndl);
       if (handleAsObjectInI18nFormat && resForObjHndl && handleAsObject && noObject.indexOf(resType) < 0 && !(isString(joinArrays) && Array.isArray(resForObjHndl))) {
-        if (!options.returnObjects && !this.options.returnObjects) {
+        if (!opt.returnObjects && !this.options.returnObjects) {
           if (!this.options.returnedObjectHandler) {
             this.logger.warn('accessing an object - but returnObjects options is not enabled!');
           }
           const r = this.options.returnedObjectHandler ? this.options.returnedObjectHandler(resUsedKey, resForObjHndl, {
-            ...options,
+            ...opt,
             ns: namespaces
           }) : `key '${key} (${this.language})' returned an object instead of string.`;
           if (returnDetails) {
             resolved.res = r;
-            resolved.usedParams = this.getUsedParamsDetails(options);
+            resolved.usedParams = this.getUsedParamsDetails(opt);
             return resolved;
           }
           return r;
@@ -604,7 +604,7 @@
               const deepKey = `${newKeyToUse}${keySeparator}${m}`;
               if (hasDefaultValue && !res) {
                 copy[m] = this.translate(deepKey, {
-                  ...options,
+                  ...opt,
                   defaultValue: shouldHandleAsObject(defaultValue) ? defaultValue[m] : undefined,
                   ...{
                     joinArrays: false,
@@ -613,7 +613,7 @@
                 });
               } else {
                 copy[m] = this.translate(deepKey, {
-                  ...options,
+                  ...opt,
                   ...{
                     joinArrays: false,
                     ns: namespaces
@@ -627,7 +627,7 @@
         }
       } else if (handleAsObjectInI18nFormat && isString(joinArrays) && Array.isArray(res)) {
         res = res.join(joinArrays);
-        if (res) res = this.extendTranslation(res, keys, options, lastKey);
+        if (res) res = this.extendTranslation(res, keys, opt, lastKey);
       } else {
         let usedDefault = false;
         let usedKey = false;
@@ -639,47 +639,47 @@
           usedKey = true;
           res = key;
         }
-        const missingKeyNoValueFallbackToKey = options.missingKeyNoValueFallbackToKey || this.options.missingKeyNoValueFallbackToKey;
+        const missingKeyNoValueFallbackToKey = opt.missingKeyNoValueFallbackToKey || this.options.missingKeyNoValueFallbackToKey;
         const resForMissing = missingKeyNoValueFallbackToKey && usedKey ? undefined : res;
         const updateMissing = hasDefaultValue && defaultValue !== res && this.options.updateMissing;
         if (usedKey || usedDefault || updateMissing) {
           this.logger.log(updateMissing ? 'updateKey' : 'missingKey', lng, namespace, key, updateMissing ? defaultValue : res);
           if (keySeparator) {
             const fk = this.resolve(key, {
-              ...options,
+              ...opt,
               keySeparator: false
             });
             if (fk && fk.res) this.logger.warn('Seems the loaded translations were in flat JSON format instead of nested. Either set keySeparator: false on init or make sure your translations are published in nested format.');
           }
           let lngs = [];
-          const fallbackLngs = this.languageUtils.getFallbackCodes(this.options.fallbackLng, options.lng || this.language);
+          const fallbackLngs = this.languageUtils.getFallbackCodes(this.options.fallbackLng, opt.lng || this.language);
           if (this.options.saveMissingTo === 'fallback' && fallbackLngs && fallbackLngs[0]) {
             for (let i = 0; i < fallbackLngs.length; i++) {
               lngs.push(fallbackLngs[i]);
             }
           } else if (this.options.saveMissingTo === 'all') {
-            lngs = this.languageUtils.toResolveHierarchy(options.lng || this.language);
+            lngs = this.languageUtils.toResolveHierarchy(opt.lng || this.language);
           } else {
-            lngs.push(options.lng || this.language);
+            lngs.push(opt.lng || this.language);
           }
           const send = (l, k, specificDefaultValue) => {
             const defaultForMissing = hasDefaultValue && specificDefaultValue !== res ? specificDefaultValue : resForMissing;
             if (this.options.missingKeyHandler) {
-              this.options.missingKeyHandler(l, namespace, k, defaultForMissing, updateMissing, options);
+              this.options.missingKeyHandler(l, namespace, k, defaultForMissing, updateMissing, opt);
             } else if (this.backendConnector?.saveMissing) {
-              this.backendConnector.saveMissing(l, namespace, k, defaultForMissing, updateMissing, options);
+              this.backendConnector.saveMissing(l, namespace, k, defaultForMissing, updateMissing, opt);
             }
             this.emit('missingKey', l, namespace, k, res);
           };
           if (this.options.saveMissing) {
             if (this.options.saveMissingPlurals && needsPluralHandling) {
               lngs.forEach(language => {
-                const suffixes = this.pluralResolver.getSuffixes(language, options);
-                if (needsZeroSuffixLookup && options[`defaultValue${this.options.pluralSeparator}zero`] && suffixes.indexOf(`${this.options.pluralSeparator}zero`) < 0) {
+                const suffixes = this.pluralResolver.getSuffixes(language, opt);
+                if (needsZeroSuffixLookup && opt[`defaultValue${this.options.pluralSeparator}zero`] && suffixes.indexOf(`${this.options.pluralSeparator}zero`) < 0) {
                   suffixes.push(`${this.options.pluralSeparator}zero`);
                 }
                 suffixes.forEach(suffix => {
-                  send([language], key + suffix, options[`defaultValue${suffix}`] || defaultValue);
+                  send([language], key + suffix, opt[`defaultValue${suffix}`] || defaultValue);
                 });
               });
             } else {
@@ -687,7 +687,7 @@
             }
           }
         }
-        res = this.extendTranslation(res, keys, options, resolved, lastKey);
+        res = this.extendTranslation(res, keys, opt, resolved, lastKey);
         if (usedKey && res === key && this.options.appendNamespaceToMissingKey) res = `${namespace}:${key}`;
         if ((usedKey || usedDefault) && this.options.parseMissingKeyHandler) {
           res = this.options.parseMissingKeyHandler(this.options.appendNamespaceToMissingKey ? `${namespace}:${key}` : key, usedDefault ? res : undefined);
@@ -695,75 +695,75 @@
       }
       if (returnDetails) {
         resolved.res = res;
-        resolved.usedParams = this.getUsedParamsDetails(options);
+        resolved.usedParams = this.getUsedParamsDetails(opt);
         return resolved;
       }
       return res;
     }
-    extendTranslation(res, key, options, resolved, lastKey) {
+    extendTranslation(res, key, opt, resolved, lastKey) {
       var _this = this;
       if (this.i18nFormat?.parse) {
         res = this.i18nFormat.parse(res, {
           ...this.options.interpolation.defaultVariables,
-          ...options
-        }, options.lng || this.language || resolved.usedLng, resolved.usedNS, resolved.usedKey, {
+          ...opt
+        }, opt.lng || this.language || resolved.usedLng, resolved.usedNS, resolved.usedKey, {
           resolved
         });
-      } else if (!options.skipInterpolation) {
-        if (options.interpolation) this.interpolator.init({
-          ...options,
+      } else if (!opt.skipInterpolation) {
+        if (opt.interpolation) this.interpolator.init({
+          ...opt,
           ...{
             interpolation: {
               ...this.options.interpolation,
-              ...options.interpolation
+              ...opt.interpolation
             }
           }
         });
-        const skipOnVariables = isString(res) && (options?.interpolation?.skipOnVariables !== undefined ? options.interpolation.skipOnVariables : this.options.interpolation.skipOnVariables);
+        const skipOnVariables = isString(res) && (opt?.interpolation?.skipOnVariables !== undefined ? opt.interpolation.skipOnVariables : this.options.interpolation.skipOnVariables);
         let nestBef;
         if (skipOnVariables) {
           const nb = res.match(this.interpolator.nestingRegexp);
           nestBef = nb && nb.length;
         }
-        let data = options.replace && !isString(options.replace) ? options.replace : options;
+        let data = opt.replace && !isString(opt.replace) ? opt.replace : opt;
         if (this.options.interpolation.defaultVariables) data = {
           ...this.options.interpolation.defaultVariables,
           ...data
         };
-        res = this.interpolator.interpolate(res, data, options.lng || this.language || resolved.usedLng, options);
+        res = this.interpolator.interpolate(res, data, opt.lng || this.language || resolved.usedLng, opt);
         if (skipOnVariables) {
           const na = res.match(this.interpolator.nestingRegexp);
           const nestAft = na && na.length;
-          if (nestBef < nestAft) options.nest = false;
+          if (nestBef < nestAft) opt.nest = false;
         }
-        if (!options.lng && resolved && resolved.res) options.lng = this.language || resolved.usedLng;
-        if (options.nest !== false) res = this.interpolator.nest(res, function () {
+        if (!opt.lng && resolved && resolved.res) opt.lng = this.language || resolved.usedLng;
+        if (opt.nest !== false) res = this.interpolator.nest(res, function () {
           for (var _len = arguments.length, args = new Array(_len), _key = 0; _key < _len; _key++) {
             args[_key] = arguments[_key];
           }
-          if (lastKey?.[0] === args[0] && !options.context) {
+          if (lastKey?.[0] === args[0] && !opt.context) {
             _this.logger.warn(`It seems you are nesting recursively key: ${args[0]} in key: ${key[0]}`);
             return null;
           }
           return _this.translate(...args, key);
-        }, options);
-        if (options.interpolation) this.interpolator.reset();
+        }, opt);
+        if (opt.interpolation) this.interpolator.reset();
       }
-      const postProcess = options.postProcess || this.options.postProcess;
+      const postProcess = opt.postProcess || this.options.postProcess;
       const postProcessorNames = isString(postProcess) ? [postProcess] : postProcess;
-      if (res != null && postProcessorNames?.length && options.applyPostProcessor !== false) {
+      if (res != null && postProcessorNames?.length && opt.applyPostProcessor !== false) {
         res = postProcessor.handle(postProcessorNames, res, key, this.options && this.options.postProcessPassResolved ? {
           i18nResolved: {
             ...resolved,
-            usedParams: this.getUsedParamsDetails(options)
+            usedParams: this.getUsedParamsDetails(opt)
           },
-          ...options
-        } : options, this);
+          ...opt
+        } : opt, this);
       }
       return res;
     }
     resolve(keys) {
-      let options = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : {};
+      let opt = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : {};
       let found;
       let usedKey;
       let exactUsedKey;
@@ -772,15 +772,15 @@
       if (isString(keys)) keys = [keys];
       keys.forEach(k => {
         if (this.isValidLookup(found)) return;
-        const extracted = this.extractFromKey(k, options);
+        const extracted = this.extractFromKey(k, opt);
         const key = extracted.key;
         usedKey = key;
         let namespaces = extracted.namespaces;
         if (this.options.fallbackNS) namespaces = namespaces.concat(this.options.fallbackNS);
-        const needsPluralHandling = options.count !== undefined && !isString(options.count);
-        const needsZeroSuffixLookup = needsPluralHandling && !options.ordinal && options.count === 0;
-        const needsContextHandling = options.context !== undefined && (isString(options.context) || typeof options.context === 'number') && options.context !== '';
-        const codes = options.lngs ? options.lngs : this.languageUtils.toResolveHierarchy(options.lng || this.language, options.fallbackLng);
+        const needsPluralHandling = opt.count !== undefined && !isString(opt.count);
+        const needsZeroSuffixLookup = needsPluralHandling && !opt.ordinal && opt.count === 0;
+        const needsContextHandling = opt.context !== undefined && (isString(opt.context) || typeof opt.context === 'number') && opt.context !== '';
+        const codes = opt.lngs ? opt.lngs : this.languageUtils.toResolveHierarchy(opt.lng || this.language, opt.fallbackLng);
         namespaces.forEach(ns => {
           if (this.isValidLookup(found)) return;
           usedNS = ns;
@@ -793,15 +793,15 @@
             usedLng = code;
             const finalKeys = [key];
             if (this.i18nFormat?.addLookupKeys) {
-              this.i18nFormat.addLookupKeys(finalKeys, key, code, ns, options);
+              this.i18nFormat.addLookupKeys(finalKeys, key, code, ns, opt);
             } else {
               let pluralSuffix;
-              if (needsPluralHandling) pluralSuffix = this.pluralResolver.getSuffix(code, options.count, options);
+              if (needsPluralHandling) pluralSuffix = this.pluralResolver.getSuffix(code, opt.count, opt);
               const zeroSuffix = `${this.options.pluralSeparator}zero`;
               const ordinalPrefix = `${this.options.pluralSeparator}ordinal${this.options.pluralSeparator}`;
               if (needsPluralHandling) {
                 finalKeys.push(key + pluralSuffix);
-                if (options.ordinal && pluralSuffix.indexOf(ordinalPrefix) === 0) {
+                if (opt.ordinal && pluralSuffix.indexOf(ordinalPrefix) === 0) {
                   finalKeys.push(key + pluralSuffix.replace(ordinalPrefix, this.options.pluralSeparator));
                 }
                 if (needsZeroSuffixLookup) {
@@ -809,11 +809,11 @@
                 }
               }
               if (needsContextHandling) {
-                const contextKey = `${key}${this.options.contextSeparator}${options.context}`;
+                const contextKey = `${key}${this.options.contextSeparator}${opt.context}`;
                 finalKeys.push(contextKey);
                 if (needsPluralHandling) {
                   finalKeys.push(contextKey + pluralSuffix);
-                  if (options.ordinal && pluralSuffix.indexOf(ordinalPrefix) === 0) {
+                  if (opt.ordinal && pluralSuffix.indexOf(ordinalPrefix) === 0) {
                     finalKeys.push(contextKey + pluralSuffix.replace(ordinalPrefix, this.options.pluralSeparator));
                   }
                   if (needsZeroSuffixLookup) {
@@ -826,7 +826,7 @@
             while (possibleKey = finalKeys.pop()) {
               if (!this.isValidLookup(found)) {
                 exactUsedKey = possibleKey;
-                found = this.getResource(code, ns, possibleKey, options);
+                found = this.getResource(code, ns, possibleKey, opt);
               }
             }
           });
@@ -938,6 +938,8 @@
       if (!found && this.options.supportedLngs) {
         codes.forEach(code => {
           if (found) return;
+          const lngScOnly = this.getScriptPartFromCode(code);
+          if (this.isSupportedCode(lngScOnly)) return found = lngScOnly;
           const lngOnly = this.getLanguagePartFromCode(code);
           if (this.isSupportedCode(lngOnly)) return found = lngOnly;
           found = this.options.supportedLngs.find(supportedLng => {
@@ -1984,11 +1986,13 @@
       };
       const done = (err, l) => {
         if (l) {
-          setLngProps(l);
-          this.translator.changeLanguage(l);
-          this.isLanguageChangingTo = undefined;
-          this.emit('languageChanged', l);
-          this.logger.log('languageChanged', l);
+          if (this.isLanguageChangingTo === lng) {
+            setLngProps(l);
+            this.translator.changeLanguage(l);
+            this.isLanguageChangingTo = undefined;
+            this.emit('languageChanged', l);
+            this.logger.log('languageChanged', l);
+          }
         } else {
           this.isLanguageChangingTo = undefined;
         }
@@ -2001,7 +2005,7 @@
       };
       const setLng = lngs => {
         if (!lng && !lngs && this.services.languageDetector) lngs = [];
-        const l = isString(lngs) ? lngs : this.services.languageUtils.getBestMatchFromCodes(lngs);
+        const l = this.services.languageUtils.getBestMatchFromCodes(isString(lngs) ? [lngs] : lngs);
         if (l) {
           if (!this.language) {
             setLngProps(l);
@@ -2029,29 +2033,29 @@
     getFixedT(lng, ns, keyPrefix) {
       var _this3 = this;
       const fixedT = function (key, opts) {
-        let options;
+        let o;
         if (typeof opts !== 'object') {
           for (var _len3 = arguments.length, rest = new Array(_len3 > 2 ? _len3 - 2 : 0), _key3 = 2; _key3 < _len3; _key3++) {
             rest[_key3 - 2] = arguments[_key3];
           }
-          options = _this3.options.overloadTranslationOptionHandler([key, opts].concat(rest));
+          o = _this3.options.overloadTranslationOptionHandler([key, opts].concat(rest));
         } else {
-          options = {
+          o = {
             ...opts
           };
         }
-        options.lng = options.lng || fixedT.lng;
-        options.lngs = options.lngs || fixedT.lngs;
-        options.ns = options.ns || fixedT.ns;
-        if (options.keyPrefix !== '') options.keyPrefix = options.keyPrefix || keyPrefix || fixedT.keyPrefix;
+        o.lng = o.lng || fixedT.lng;
+        o.lngs = o.lngs || fixedT.lngs;
+        o.ns = o.ns || fixedT.ns;
+        if (o.keyPrefix !== '') o.keyPrefix = o.keyPrefix || keyPrefix || fixedT.keyPrefix;
         const keySeparator = _this3.options.keySeparator || '.';
         let resultKey;
-        if (options.keyPrefix && Array.isArray(key)) {
-          resultKey = key.map(k => `${options.keyPrefix}${keySeparator}${k}`);
+        if (o.keyPrefix && Array.isArray(key)) {
+          resultKey = key.map(k => `${o.keyPrefix}${keySeparator}${k}`);
         } else {
-          resultKey = options.keyPrefix ? `${options.keyPrefix}${keySeparator}${key}` : key;
+          resultKey = o.keyPrefix ? `${o.keyPrefix}${keySeparator}${key}` : key;
         }
-        return _this3.t(resultKey, options);
+        return _this3.t(resultKey, o);
       };
       if (isString(lng)) {
         fixedT.lng = lng;